mkhexgrid 0.1.1
The Makefile has the compiler and prefix hardcoded and needs to be
changed before building. During installation, it also tries to change
owner and group to 0. In Mac OS 10.9, there's also a namescape
conflict for exceptions. The Formula handles these issues for you.

The Formula depends on GD with FreeType enabled. This can be a problem
if you have GD without FreeType installed.

Closes #95.

Signed-off-by: Misty De Meo <mistydemeo@gmail.com>